Split structure and content validation
