+++ src/ingest_validation_tools/directory_structure_validator.py
@@ -1,44 +1,9 @@
 from os import walk
 import re
 from fnmatch import fnmatch
-from importlib import import_module
 
 
-def default_factory(base_path, assay_type):
-    """
-    default validator factory
-    """
-    return DefaultDirectoryContentValidator(base_path, assay_type)
-
-
-DIRECTORY_CONTENT_VALIDATOR_FACTORY = default_factory
-
-
-class DirectoryContentValidator(object):
-    def __init__(self, base_path, assay_type):
-        self.base_path = base_path
-        self.assay_type = assay_type
-
-    def collect_errors(self):
-        """
-        returns string or None.  Multiple errors are to be reported
-        via a single multi-line string.
-        """
-        return None
-
-
-class DefaultDirectoryContentValidator(DirectoryContentValidator):
-    """
-    Implements backward compatible calls to content validation modules
-    """
-
-    def collect_errors(self):
-        validator = import_module(
-            f'ingest_validation_tools.content_validation.{self.assay_type}')
-        return validator.collect_errors(self.base_path)
-
-
-class DirectoryValidationErrors(Exception):
+class DirectoryStructureValidationErrors(Exception):
     def __init__(self, errors):
         self.errors = errors
 
@@ -51,7 +16,7 @@ def validate_directory_structure(path, paths_dict, dataset_ignore_globs=[]):
     required_patterns, allowed_patterns = _get_required_allowed(paths_dict)
     required_missing_errors, not_allowed_errors = ([], [])
     if not path.exists():
-        raise DirectoryValidationErrors({
+        raise DirectoryStructureValidationErrors({
             'No such file or directory': str(path)
         })
     actual_paths = []
@@ -81,7 +46,7 @@ def validate_directory_structure(path, paths_dict, dataset_ignore_globs=[]):
     if required_missing_errors:
         errors['Required but missing'] = required_missing_errors
     if errors:
-        raise DirectoryValidationErrors(errors)
+        raise DirectoryStructureValidationErrors(errors)
 
 
 def _get_required_allowed(dir_schema):
@@ -103,10 +68,3 @@ def _get_required_allowed(dir_schema):
         item['pattern'] for item in dir_schema
         if 'required' not in item or item['required']]
     return required, allowed
-
-
-def validate_directory_content(assay_type, data_path):
-    validator = DIRECTORY_CONTENT_VALIDATOR_FACTORY(data_path, assay_type)
-    errors = validator.collect_errors()
-    if errors:
-        raise DirectoryValidationErrors(errors)
